30 tests passing, buy item currently works perfectly
